The general idea is solid, but...
Hi Brendan,

Your proposal is clear but you are missing a series of storyboards for the interface.

Please spend time here breaking down how the gameplay will work in a precise way.

Show the various screens that a user can expect to encounter.

Giving these various screens thought will help you to decompose the problem (identify where to start with your implementation) a lot more easily.

Please make edits and resubmit ASAP.

As it stands this is at best a Level 2 proposal. 

Thanks and talk to you soon,

Mr. Gordon